这个版本已经比较完整了，各种bug得到了解决，deflect和drop的概率也都已经算上去了，产生数据包的时间间隔变成了随机情况。接下去就是先对random和flooding的实验进行对比，然后先进行会议论文的投稿？

+++ src/nanonetworks/model/q-routing.h
@@ -48,7 +48,7 @@ public:
 	//uint32_t LookupPreNode();the algorithm did not need to lookup some node, these will be finished at upper layer.
 	void AddPreNodeNew(uint32_t distId, uint32_t nextID, uint32_t Qvalue,
 			uint32_t HopCount);
-	void UpdatePreNode(uint32_t dstId, uint32_t nextId, uint32_t reward,
+	void UpdatePreNode(uint32_t dstId, uint32_t nextId, double reward,
 			uint32_t HopCount);
 
 	void SetRouteUn(uint32_t nextId);
